requirements.txt
google-cloud-resource-manager
wheel
google-resumable-media
google-cloud-storage==2.17.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-logging
google-auth-oauthlib==1.2.1
google-cloud-bigquery
azure-identity==1.17.1
azure-storage-blob==12.21.0
oauth2client
backoff
aiohttp
pytz
pandas
db-dtypes
pydantic==2.9.1
pyyaml
humanfriendly
responses
gitpython
git+https://github.com/broadinstitute/ops_util_module.git@3ceee50aa266f950f1747406ce02e21a6ea59307#egg=ops_util_module
